// src/app/student&staff/model/student.model.ts
export class Student {
        id?: number;
        regno?: number;
        rollno?: number;
        class?: string;
        name?: string;
        fname?: string;
        mname?: string;
        dob?: Date;
        dor?: Date;
        address?: string;
        city?: string;
        stat?: string;
        pin?: string;
        phone?: string;        
}